# shared/schemas/api.openapi.yaml
openapi: 3.0.3
info:
  title: Family Trivia API
  description: Backend API for the Family Trivia game
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local SAM development
paths:
...
  /admin/family-feud:
    get:
      summary: Admin — inspect family feud state
      operationId: adminGetFamilyFeud
      tags: [Admin]
      parameters:
        - { name: adminUserId, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Feud state for debugging
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Admin — reset or delete family feud round
      operationId: adminResetFamilyFeud
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adminUserId, action]
              properties:
                adminUserId:
                  type: string
                action:
                  type: string
                  enum: [reset-current, delete-round, reset-rotation]
                roundId:
                  type: string
      responses:
        "200":
          description: Action completed
          content:
            application/json:
              schema:
                type: object
...